backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pandas>=2.0.0
openai>=1.10.0
azure-ai-formrecognizer>=3.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
openpyxl>=3.1.0
tenacity>=8.0.0
sqlalchemy>=2.0.0
